api(views): modulariza views em uma pasta separada
